Move old repository code to legacy package
