+++ multipacks-cli/src/main/java/multipacks/cli/Main.java
@@ -24,7 +24,7 @@ import java.util.stream.Stream;
 
 import multipacks.bundling.BundleIgnore;
 import multipacks.bundling.BundleInclude;
-import multipacks.management.PacksRepository;
+import multipacks.management.legacy.PacksRepository;
 import multipacks.packs.PackIdentifier;
 import multipacks.plugins.MultipacksDefaultPlugin;
 import multipacks.plugins.MultipacksPlugin;
