+++ multipacks-cli/src/main/java/multipacks/cli/MultipacksCLI.java
@@ -41,9 +41,9 @@ import multipacks.bundling.BundleIgnore;
 import multipacks.bundling.BundleInclude;
 import multipacks.bundling.PackBundler;
 import multipacks.bundling.PackagingFailException;
-import multipacks.management.LocalRepository;
-import multipacks.management.PacksRepository;
-import multipacks.management.PacksUploadable;
+import multipacks.management.legacy.LocalRepository;
+import multipacks.management.legacy.PacksRepository;
+import multipacks.management.legacy.PacksUploadable;
 import multipacks.packs.Pack;
 import multipacks.packs.PackIdentifier;
 import multipacks.packs.PackIndex;
