+++ multipacks-engine/src/main/java/multipacks/management/legacy/LocalRepository.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.management;
+package multipacks.management.legacy;
 
 import java.io.File;
 import java.io.IOException;
