+++ multipacks-engine/src/main/java/multipacks/management/legacy/PacksUploadable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.management;
+package multipacks.management.legacy;
 
 import multipacks.packs.Pack;
